+++ HISTORY.rst
@@ -10,7 +10,13 @@ TODO
 
 * ``Domain.users`` should returns ``User`` objects
 * update submission status after reload a just finalized submission
+* update submission status after create a new submission
+* update submission status after ``get_submission_by_name``
 
+Features
+^^^^^^^^
+
+* test for an empty submission (no samples)
 
 0.2.0 (2018-10-23)
 ------------------